+++ generators/graph/voxel_generator_graph.cpp
@@ -1812,6 +1812,90 @@ void VoxelGeneratorGraph::get_configuration_warnings(TypedArray<String> &out_war
 	}
 }
 
+// Gets a hash of a given object from its properties. If properties are objects too, they are recursively parsed.
+// Note that restricting to editable properties is important to avoid costly properties with objects such as textures or
+// meshes
+static uint64_t get_deep_hash(const Object &obj,
+		uint32_t property_usage = PROPERTY_USAGE_STORAGE | PROPERTY_USAGE_EDITOR, uint64_t hash = 0) {
+	ZN_PROFILE_SCOPE();
+
+	hash = hash_djb2_one_64(obj.get_class_name().hash(), hash);
+
+	List<PropertyInfo> properties;
+	obj.get_property_list(&properties, false);
+
+	// I'd like to use ConstIterator since I only read that list but that isn't possible :shrug:
+	for (List<PropertyInfo>::Iterator it = properties.begin(); it != properties.end(); ++it) {
+		const PropertyInfo property = *it;
+
+		if ((property.usage & property_usage) != 0) {
+			const Variant value = obj.get(property.name);
+			uint64_t value_hash = 0;
+
+			if (value.get_type() == Variant::OBJECT) {
+				const Object *obj_value = value.operator Object *();
+				if (obj_value != nullptr) {
+					value_hash = get_deep_hash(*obj_value, hash, property_usage);
+				}
+
+			} else {
+				value_hash = value.hash();
+			}
+
+			hash = hash_djb2_one_64(value_hash, hash);
+		}
+	}
+
+	return hash;
+}
+
+uint64_t VoxelGeneratorGraph::get_output_graph_hash() const {
+	const VoxelGraphNodeDB &type_db = VoxelGraphNodeDB::get_singleton();
+	std::vector<uint32_t> terminal_nodes;
+
+	// Not using the generic `get_terminal_nodes` function because our terminal nodes do have outputs
+	_graph.for_each_node_const([&terminal_nodes, &type_db](const ProgramGraph::Node &node) {
+		const VoxelGraphNodeDB::NodeType &type = type_db.get_type(node.type_id);
+		if (type.category == VoxelGraphNodeDB::CATEGORY_OUTPUT) {
+			terminal_nodes.push_back(node.id);
+		}
+	});
+
+	// Sort for determinism
+	std::sort(terminal_nodes.begin(), terminal_nodes.end());
+
+	std::vector<uint32_t> order;
+	_graph.find_dependencies(terminal_nodes, order);
+
+	std::vector<uint64_t> node_hashes;
+	uint64_t hash = hash_djb2_one_64(0);
+
+	// Connections should be implicitely hashed due to the order of iteration
+	for (uint32_t node_id : order) {
+		ProgramGraph::Node &node = _graph.get_node(node_id);
+		hash = hash_djb2_one_64(node.type_id, hash);
+
+		for (const Variant &v : node.params) {
+			if (v.get_type() == Variant::OBJECT) {
+				const Object *obj = v.operator Object *();
+				if (obj != nullptr) {
+					// Note, the obtained hash can change here even if the result is identical, because it's hard to
+					// tell which properties contribute to the result. This should be rare though.
+					hash = hash_djb2_one_64(get_deep_hash(*obj), hash);
+				}
+			} else {
+				hash = hash_djb2_one_64(v.hash(), hash);
+			}
+		}
+
+		for (const Variant &v : node.default_inputs) {
+			hash = hash_djb2_one_float_64(v.hash(), hash);
+		}
+	}
+
+	return hash;
+}
+
 #endif // TOOLS_ENABLED
 
 // Binding land