+++ util/tasks/threaded_task_runner.cpp
@@ -78,15 +78,16 @@ void ThreadedTaskRunner::set_batch_count(uint32_t count) {
 }
 
 void ThreadedTaskRunner::set_priority_update_period(uint32_t milliseconds) {
-	_priority_update_period = milliseconds;
+	_priority_update_period_ms = milliseconds;
 }
 
-void ThreadedTaskRunner::enqueue(IThreadedTask *task) {
+void ThreadedTaskRunner::enqueue(IThreadedTask *task, bool serial) {
 	ZN_ASSERT(task != nullptr);
+	TaskItem t;
+	t.task = task;
+	t.is_serial = serial;
 	{
 		MutexLock lock(_tasks_mutex);
-		TaskItem t;
-		t.task = task;
 		_tasks.push_back(t);
 		++_debug_received_tasks;
 	}
@@ -94,7 +95,7 @@ void ThreadedTaskRunner::enqueue(IThreadedTask *task) {
 	_tasks_semaphore.post();
 }
 
-void ThreadedTaskRunner::enqueue(Span<IThreadedTask *> new_tasks) {
+void ThreadedTaskRunner::enqueue(Span<IThreadedTask *> new_tasks, bool serial) {
 #ifdef DEBUG_ENABLED
 	for (size_t i = 0; i < new_tasks.size(); ++i) {
 		ZN_ASSERT(new_tasks[i] != nullptr);
@@ -107,6 +108,7 @@ void ThreadedTaskRunner::enqueue(Span<IThreadedTask *> new_tasks) {
 		for (size_t i = 0; i < new_tasks.size(); ++i) {
 			TaskItem t;
 			t.task = new_tasks[i];
+			t.is_serial = serial;
 			_tasks[dst_begin + i] = t;
 		}
 		_debug_received_tasks += new_tasks.size();
@@ -139,54 +141,86 @@ void ThreadedTaskRunner::thread_func(ThreadData &data) {
 	std::vector<IThreadedTask *> cancelled_tasks;
 
 	while (!data.stop) {
+		bool is_running_serial_task = false;
+		bool task_queue_was_empty = false;
 		{
 			ZN_PROFILE_SCOPE_NAMED("Task pickup");
 
 			data.debug_state = STATE_PICKING;
 			const uint64_t now = Time::get_singleton()->get_ticks_msec();
 
-			MutexLock lock(_tasks_mutex);
+			{
+				MutexLock lock(_tasks_mutex);
 
-			// Pick best tasks
-			for (uint32_t bi = 0; bi < _batch_count && _tasks.size() != 0; ++bi) {
-				size_t best_index = 0; // Take first by default, this is a valid index
-				int best_priority = 999999;
+				// Pick best tasks
+				for (uint32_t bi = 0; bi < _batch_count && _tasks.size() != 0; ++bi) {
+					size_t best_index = 0; // Take first by default, this is a valid index
+					int best_priority = 999999;
+					bool picked_task = false;
 
-				// TODO This takes a lot of time when there are many queued tasks. Use a better container?
-				for (size_t i = 0; i < _tasks.size(); ++i) {
-					TaskItem &item = _tasks[i];
-					ZN_ASSERT(item.task != nullptr);
+					// Find best task to pick
+					// TODO Optimize: this takes a lot of time when there are many queued tasks. Use a better container?
+					for (size_t i = 0; i < _tasks.size();) {
+						TaskItem &item = _tasks[i];
+						ZN_ASSERT(item.task != nullptr);
 
-					if (now - item.last_priority_update_time_ms > _priority_update_period) {
-						// Calling `get_priority()` first since it can update cancellation
-						// (not clear API tho, might review that in the future)
-						item.cached_priority = item.task->get_priority();
+						// Process priority and cancellation
+						if (now - item.last_priority_update_time_ms > _priority_update_period_ms) {
+							// Calling `get_priority()` first since it can update cancellation
+							// (not clear API tho, might review that in the future)
+							item.cached_priority = item.task->get_priority();
 
-						if (item.task->is_cancelled()) {
-							cancelled_tasks.push_back(item.task);
-							_tasks[i] = _tasks.back();
-							_tasks.pop_back();
-							--i;
-							continue;
+							if (item.task->is_cancelled()) {
+								cancelled_tasks.push_back(item.task);
+								_tasks[i] = _tasks.back();
+								_tasks.pop_back();
+								continue;
+							}
+
+							item.last_priority_update_time_ms = now;
 						}
 
-						item.last_priority_update_time_ms = now;
+						// Pick item if it has better priority.
+						// If the item is serial, there must not be a serial task already running.
+						if (item.cached_priority < best_priority && !(item.is_serial && _is_serial_task_running)) {
+							best_priority = item.cached_priority;
+							// This index should remain valid even if some tasks are removed because the "remove and
+							// swap back" technique only affects items coming after
+							best_index = i;
+							picked_task = true;
+						}
+
+						++i;
 					}
 
-					if (item.cached_priority < best_priority) {
-						best_priority = item.cached_priority;
-						best_index = i;
+					if (picked_task) {
+						tasks.push_back(_tasks[best_index]);
+						ZN_ASSERT(best_index < _tasks.size());
+						_tasks[best_index] = _tasks.back();
+						_tasks.pop_back();
+					}
+
+				} // For each task to pick
+
+				// If we picked up a serial task, we must set the shared boolean to `true`.
+				// More than one serial task can be in the list of tasks the current thread picks up,
+				// so we update the boolean after picking them all.
+				// This must be the only place it can be set to `true`, and is guarded by mutex.
+				if (_is_serial_task_running == false) { // Only an optimization, this doesnt actually do thread-safety
+					for (unsigned int i = 0; i < tasks.size(); ++i) {
+						if (tasks[i].is_serial) {
+							// Write to member var so all threads can check this
+							_is_serial_task_running = true;
+							// Write to thread-local variable so we know it is the current thread
+							is_running_serial_task = true;
+							break;
+						}
 					}
 				}
 
-				// If not all tasks were cancelled
-				if (_tasks.size() != 0) {
-					tasks.push_back(_tasks[best_index]);
-					ZN_ASSERT(best_index < _tasks.size());
-					_tasks[best_index] = _tasks.back();
-					_tasks.pop_back();
-				}
-			}
+				task_queue_was_empty = _tasks.size() == 0;
+
+			} // Tasks queue mutex lock
 		}
 
 		if (cancelled_tasks.size() > 0) {
@@ -195,22 +229,35 @@ void ThreadedTaskRunner::thread_func(ThreadData &data) {
 				_completed_tasks.push_back(cancelled_tasks[i]);
 				++_debug_completed_tasks;
 			}
+			cancelled_tasks.clear();
 		}
-		cancelled_tasks.clear();
 
 		//print_line(String("Processing {0} tasks").format(varray(tasks.size())));
 
 		if (tasks.empty()) {
-			data.debug_state = STATE_WAITING;
+			if (task_queue_was_empty) {
+				// The task queue is empty, will wait until more tasks are posted.
+				// If a task is posted between the moment we last checked the queue and now,
+				// the semaphore will have one count to decrement and we'll not stop here.
 
-			// Wait for more tasks
-			data.waiting = true;
-			_tasks_semaphore.wait();
-			data.waiting = false;
+				data.debug_state = STATE_WAITING;
+
+				// Wait for more tasks
+				data.waiting = true;
+				_tasks_semaphore.wait();
+				data.waiting = false;
+
+			} else {
+				// The current thread was not allowed to pick tasks currently in the queue (they could be serial and one
+				// is already running). So we'll wait for a very short time before retrying.
+				// (alternative would be to post the semaphore after each serial task?)
+				Thread::sleep_usec(1000);
+			}
 
 		} else {
 			data.debug_state = STATE_RUNNING;
 
+			// Run each task
 			for (size_t i = 0; i < tasks.size(); ++i) {
 				TaskItem &item = tasks[i];
 				if (!item.task->is_cancelled()) {
@@ -219,6 +266,16 @@ void ThreadedTaskRunner::thread_func(ThreadData &data) {
 					item.task->run(ctx);
 				}
 			}
+
+			// If the current thread just ran serial tasks
+			if (is_running_serial_task) {
+				ZN_ASSERT(_is_serial_task_running);
+				// Reset back the boolean so any thread can pick serial tasks now.
+				// This is the only place we set it to `false`, and can only be `true` already when that happens,
+				// so locking the mutex should not be necessary.
+				_is_serial_task_running = false;
+			}
+
 			{
 				MutexLock lock(_completed_tasks_mutex);
 				for (size_t i = 0; i < tasks.size(); ++i) {
@@ -236,7 +293,7 @@ void ThreadedTaskRunner::thread_func(ThreadData &data) {
 }
 
 void ThreadedTaskRunner::wait_for_all_tasks() {
-	const uint32_t suspicious_delay_msec = 10000;
+	const uint32_t suspicious_delay_msec = 10'000;
 
 	uint32_t before = Time::get_singleton()->get_ticks_msec();
 	bool error1_reported = false;